backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
celery
redis
pdfplumber
python-docx
pdf2image
pytesseract
Pillow
aiofiles
regex
python-multipart
requests
motor